wip - CTE comment decoding
